🌱 refactor: split oversized test files (batch 2: MCP/config/AI)
Split 5 test files (1253-1327 lines each) into focused modules:
- config.test.ts → 3 files (basic, agent-sse, fallback-modes)
- useAIPredictions.test.ts → 2 files (hook, utils)
- useMetricsHistory.test.ts → 3 files (core, trends, advanced)
- useCachedData.rest-paths.test.ts → 3 files (fetchers, sse-agent, kubectl)
- useCertManager.test.ts → 3 files (basic, detection, caching)
