+++ web/src/hooks/__tests__/useCachedData-fetchers.test.ts
@@ -183,14 +183,6 @@ describe('useCachedData', () => {
     mod = await import('../useCachedData')
     return mod
   }
-
-  // ========================================================================
-  // useCachedPods
-  // ========================================================================
-
-  // ========================================================================
-  // Warning events fetcher
-  // ========================================================================
   describe('useCachedWarningEvents fetcher paths', () => {
     afterEach(() => { vi.unstubAllGlobals() })
 
@@ -779,511 +771,4 @@ describe('useCachedData', () => {
   // ========================================================================
   // coreFetchers remaining edge cases
   // ========================================================================
-  describe('coreFetchers — edge cases', () => {
-    afterEach(() => { vi.unstubAllGlobals() })
-
-    it('coreFetchers.deploymentIssues returns empty when both unavailable', async () => {
-      mockIsAgentUnavailable.mockReturnValue(true)
-      mockIsBackendUnavailable.mockReturnValue(true)
-
-      const { coreFetchers } = await loadModule()
-      const issues = await coreFetchers.deploymentIssues()
-      expect(issues).toEqual([])
-    })
-
-    it('coreFetchers.deployments returns empty when both unavailable', async () => {
-      mockIsAgentUnavailable.mockReturnValue(true)
-      mockIsBackendUnavailable.mockReturnValue(true)
-
-      const { coreFetchers } = await loadModule()
-      const deps = await coreFetchers.deployments()
-      expect(deps).toEqual([])
-    })
-
-    it('coreFetchers.securityIssues uses agent kubectl when available', async () => {
-      mockClusterCacheRef.clusters = [{ name: 'c1', reachable: true }] as typeof mockClusterCacheRef.clusters
-      mockIsAgentUnavailable.mockReturnValue(false)
-      mockKubectlProxy.exec.mockResolvedValue({
-        exitCode: 0,
-        output: JSON.stringify({
-          items: [{
-            metadata: { name: 'priv-pod', namespace: 'default' },
-            spec: {
-              containers: [{ securityContext: { privileged: true } }],
-            },
-          }],
-        }),
-      })
-
-      const { coreFetchers } = await loadModule()
-      const issues = await coreFetchers.securityIssues()
-      expect(issues.length).toBeGreaterThan(0)
-    })
-
-    it('coreFetchers.securityIssues returns empty when all unavailable', async () => {
-      mockIsAgentUnavailable.mockReturnValue(true)
-      mockIsBackendUnavailable.mockReturnValue(true)
-
-      const { coreFetchers } = await loadModule()
-      const issues = await coreFetchers.securityIssues()
-      expect(issues).toEqual([])
-    })
-  })
-
-  // ========================================================================
-  // NEW: Deep branch coverage — SSE streaming paths
-  // ========================================================================
-  describe('SSE streaming — onClusterData accumulation and catch-fallback', () => {
-    afterEach(() => { vi.unstubAllGlobals() })
-
-    it.skip('fetchViaSSE accumulates data across multiple onClusterData calls', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      // SSE delivers data from two clusters via onClusterData
-      mockFetchSSE.mockImplementation(async (opts: { onClusterData: (c: string, items: unknown[]) => void }) => {
-        opts.onClusterData('c1', [{ name: 'svc-a' }, { name: 'svc-b' }])
-        opts.onClusterData('c2', [{ name: 'svc-c' }])
-        return [{ name: 'svc-a' }, { name: 'svc-b' }, { name: 'svc-c' }]
-      })
-
-      const { useCachedServices } = await loadModule()
-      useCachedServices()
-
-      const progressiveFetcher = capturedOpts.progressiveFetcher as (onProgress: (p: unknown[]) => void) => Promise<unknown[]>
-      const onProgress = vi.fn()
-      const result = await progressiveFetcher(onProgress)
-
-      // Three total items from two clusters
-      expect(result).toHaveLength(3)
-    })
-
-    it('fetchViaSSE catches SSE error and falls back to fetchFromAllClusters', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      // SSE throws an error
-      mockFetchSSE.mockRejectedValue(new Error('EventSource connection refused'))
-
-      // REST fallback: fetchFromAllClusters needs clusters
-      mockClusterCacheRef.clusters = [{ name: 'c1', reachable: true }] as typeof mockClusterCacheRef.clusters
-
-      // REST per-cluster response
-      vi.stubGlobal('fetch', vi.fn().mockResolvedValue({
-        ok: true,
-        text: vi.fn().mockResolvedValue(JSON.stringify({ pvcs: [{ name: 'rest-pvc' }] })),
-      }))
-
-      const { useCachedPVCs } = await loadModule()
-      useCachedPVCs()
-
-      const progressiveFetcher = capturedOpts.progressiveFetcher as (onProgress: (p: unknown[]) => void) => Promise<unknown[]>
-      const result = await progressiveFetcher(vi.fn())
-
-      // Should have fallen back to REST and gotten data
-      expect(result).toHaveLength(1)
-      expect(result[0]).toHaveProperty('name', 'rest-pvc')
-    })
-
-    it.skip('fetchViaSSE calls onProgress during progressive accumulation for pods', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      mockFetchSSE.mockImplementation(async (opts: { onClusterData: (c: string, items: unknown[]) => void }) => {
-        opts.onClusterData('c1', [{ name: 'pod-1', restarts: 5 }])
-        opts.onClusterData('c2', [{ name: 'pod-2', restarts: 0 }])
-        return [{ name: 'pod-1', restarts: 5 }, { name: 'pod-2', restarts: 0 }]
-      })
-
-      const { useCachedPods } = await loadModule()
-      useCachedPods()
-
-      const progressiveFetcher = capturedOpts.progressiveFetcher as (onProgress: (p: unknown[]) => void) => Promise<unknown[]>
-      const onProgress = vi.fn()
-      const result = await progressiveFetcher(onProgress)
-
-      // Should sort by restarts desc and slice to limit
-      expect(result[0]).toHaveProperty('restarts', 5)
-      expect(result[1]).toHaveProperty('restarts', 0)
-    })
-  })
-
-  // ========================================================================
-  // NEW: Agent fallback chains — fetchDeploymentsViaAgent edge cases
-  // ========================================================================
-  describe('agent fallback chains — fetchDeploymentsViaAgent', () => {
-    afterEach(() => { vi.unstubAllGlobals() })
-
-    it('fetchDeploymentsViaAgent returns empty when agent is unavailable', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      mockClusterCacheRef.clusters = [{ name: 'c1', context: 'c1-ctx', reachable: true }] as typeof mockClusterCacheRef.clusters
-      mockIsAgentUnavailable.mockReturnValue(true)
-      mockIsBackendUnavailable.mockReturnValue(true)
-
-      const { useCachedDeployments } = await loadModule()
-      useCachedDeployments()
-
-      const fetcher = capturedOpts.fetcher as () => Promise<unknown[]>
-      await expect(fetcher()).rejects.toThrow('No data source available')
-    })
-
-    it('fetchDeploymentsViaAgent handles agent JSON returning null for each cluster', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      mockClusterCacheRef.clusters = [{ name: 'c1', context: 'c1-ctx', reachable: true }] as typeof mockClusterCacheRef.clusters
-      mockIsAgentUnavailable.mockReturnValue(false)
-
-      // Agent returns ok but JSON fails (returns null via .catch)
-      vi.stubGlobal('fetch', vi.fn().mockResolvedValue({
-        ok: true,
-        json: vi.fn().mockResolvedValue(null),
-      }))
-
-      const { useCachedDeployments } = await loadModule()
-      useCachedDeployments() // no cluster => uses fetchDeploymentsViaAgent
-
-      const fetcher = capturedOpts.fetcher as () => Promise<unknown[]>
-      const result = await fetcher()
-
-      // fetchDeploymentsViaAgent: null data => throws 'Invalid JSON'
-      // settledWithConcurrency settles, accumulated is empty, returns []
-      expect(Array.isArray(result)).toBe(true)
-    })
-
-    it('fetchDeploymentsViaAgent tags results with short cluster name, not context', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      mockClusterCacheRef.clusters = [{ name: 'prod', context: 'default/api-server:6443/admin', reachable: true }] as typeof mockClusterCacheRef.clusters
-      mockIsAgentUnavailable.mockReturnValue(false)
-
-      vi.stubGlobal('fetch', vi.fn().mockResolvedValue({
-        ok: true,
-        json: vi.fn().mockResolvedValue({
-          deployments: [{ name: 'dep-1', namespace: 'default', cluster: 'default/api-server:6443/admin' }],
-        }),
-      }))
-
-      const { useCachedDeployments } = await loadModule()
-      useCachedDeployments()
-
-      const fetcher = capturedOpts.fetcher as () => Promise<Array<{ cluster: string }>>
-      const result = await fetcher()
-
-      // Should use short name 'prod', not the context path
-      expect(result[0].cluster).toBe('prod')
-    })
-  })
-
-  // ========================================================================
-  // NEW: fetchWorkloadsFromAgent edge cases
-  // ========================================================================
-  describe('fetchWorkloadsFromAgent edge cases', () => {
-    afterEach(() => { vi.unstubAllGlobals() })
-
-    it('returns null when agent has no clusters', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      mockClusterCacheRef.clusters = [] as typeof mockClusterCacheRef.clusters
-      mockIsAgentUnavailable.mockReturnValue(false)
-      mockIsBackendUnavailable.mockReturnValue(true)
-
-      const { useCachedWorkloads } = await loadModule()
-      useCachedWorkloads()
-
-      const fetcher = capturedOpts.fetcher as () => Promise<unknown[]>
-      const result = await fetcher()
-      // No clusters => null from agent => falls through, backend unavailable => empty
-      expect(result).toEqual([])
-    })
-
-    it('returns null when agent fetch fails for all clusters', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      mockClusterCacheRef.clusters = [{ name: 'c1', context: 'c1-ctx', reachable: true }] as typeof mockClusterCacheRef.clusters
-      mockIsAgentUnavailable.mockReturnValue(false)
-      mockIsBackendUnavailable.mockReturnValue(true)
-
-      // Agent fetch throws
-      vi.stubGlobal('fetch', vi.fn().mockRejectedValue(new Error('Network error')))
-
-      const { useCachedWorkloads } = await loadModule()
-      useCachedWorkloads()
-
-      const fetcher = capturedOpts.fetcher as () => Promise<unknown[]>
-      const result = await fetcher()
-
-      // All cluster fetches fail => accumulated is empty => returns null => backend unavailable => []
-      expect(result).toEqual([])
-    })
-
-    it('progressive fetcher for workloads calls onProgress', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      mockClusterCacheRef.clusters = [{ name: 'c1', context: 'c1-ctx', reachable: true }] as typeof mockClusterCacheRef.clusters
-      mockIsAgentUnavailable.mockReturnValue(false)
-
-      vi.stubGlobal('fetch', vi.fn().mockResolvedValue({
-        ok: true,
-        json: vi.fn().mockResolvedValue({
-          deployments: [{ name: 'wl-1', status: 'running', replicas: 1, readyReplicas: 1 }],
-        }),
-      }))
-
-      const { useCachedWorkloads } = await loadModule()
-      useCachedWorkloads()
-
-      const progressiveFetcher = capturedOpts.progressiveFetcher as (onProgress: (p: unknown[]) => void) => Promise<unknown[]>
-      const onProgress = vi.fn()
-      const result = await progressiveFetcher(onProgress)
-
-      expect(result).not.toBeNull()
-      expect(onProgress).toHaveBeenCalled()
-    })
-  })
-
-  // ========================================================================
-  // NEW: Security scanning via kubectl — additional branch coverage
-  // ========================================================================
-  describe('security scanning via kubectl — additional branches', () => {
-    afterEach(() => { vi.unstubAllGlobals() })
-
-    it('detects host PID and host IPC in separate pods', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      mockClusterCacheRef.clusters = [{ name: 'prod', context: 'prod-ctx', reachable: true }] as typeof mockClusterCacheRef.clusters
-      mockIsAgentUnavailable.mockReturnValue(false)
-
-      mockKubectlProxy.exec.mockResolvedValue({
-        exitCode: 0,
-        output: JSON.stringify({
-          items: [
-            {
-              metadata: { name: 'pid-pod', namespace: 'system' },
-              spec: {
-                hostPID: true,
-                containers: [{ securityContext: { runAsNonRoot: true } }],
-              },
-            },
-            {
-              metadata: { name: 'ipc-pod', namespace: 'system' },
-              spec: {
-                hostIPC: true,
-                containers: [{ securityContext: { runAsNonRoot: true } }],
-              },
-            },
-          ],
-        }),
-      })
-
-      const { useCachedSecurityIssues } = await loadModule()
-      useCachedSecurityIssues()
-
-      const fetcher = capturedOpts.fetcher as () => Promise<Array<{ name: string; issue: string; severity: string }>>
-      const issues = await fetcher()
-
-      expect(issues.some(i => i.name === 'pid-pod' && i.issue === 'Host PID enabled')).toBe(true)
-      expect(issues.some(i => i.name === 'ipc-pod' && i.issue === 'Host IPC enabled')).toBe(true)
-    })
-
-    it('detects capabilities added without dropping any', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      mockClusterCacheRef.clusters = [{ name: 'prod', context: 'prod-ctx', reachable: true }] as typeof mockClusterCacheRef.clusters
-      mockIsAgentUnavailable.mockReturnValue(false)
-
-      mockKubectlProxy.exec.mockResolvedValue({
-        exitCode: 0,
-        output: JSON.stringify({
-          items: [
-            {
-              metadata: { name: 'cap-pod', namespace: 'apps' },
-              spec: {
-                containers: [
-                  {
-                    securityContext: {
-                      capabilities: { add: ['SYS_ADMIN', 'NET_ADMIN'] },
-                    },
-                  },
-                ],
-              },
-            },
-          ],
-        }),
-      })
-
-      const { useCachedSecurityIssues } = await loadModule()
-      useCachedSecurityIssues()
-
-      const fetcher = capturedOpts.fetcher as () => Promise<Array<{ issue: string }>>
-      const issues = await fetcher()
-      expect(issues.some(i => i.issue === 'Capabilities not dropped')).toBe(true)
-    })
-
-    it.skip('does NOT flag capabilities when caps are properly dropped', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      mockClusterCacheRef.clusters = [{ name: 'prod', context: 'prod-ctx', reachable: true }] as typeof mockClusterCacheRef.clusters
-      mockIsAgentUnavailable.mockReturnValue(false)
-
-      mockKubectlProxy.exec.mockResolvedValue({
-        exitCode: 0,
-        output: JSON.stringify({
-          items: [
-            {
-              metadata: { name: 'good-pod', namespace: 'secure' },
-              spec: {
-                containers: [
-                  {
-                    securityContext: {
-                      runAsNonRoot: true,
-                      readOnlyRootFilesystem: true,
-                      capabilities: { drop: ['ALL'], add: ['NET_BIND_SERVICE'] },
-                    },
-                  },
-                ],
-              },
-            },
-          ],
-        }),
-      })
-
-      const { useCachedSecurityIssues } = await loadModule()
-      useCachedSecurityIssues()
-
-      const fetcher = capturedOpts.fetcher as () => Promise<Array<{ issue: string }>>
-      const issues = await fetcher()
-      expect(issues.some(i => i.issue === 'Capabilities not dropped')).toBe(false)
-    })
-
-    it('filters by specific cluster when cluster arg provided', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      mockClusterCacheRef.clusters = [ { name: 'prod', context: 'prod-ctx', reachable: true }, { name: 'staging', context: 'staging-ctx', reachable: true }, ] as typeof mockClusterCacheRef.clusters
-      mockIsAgentUnavailable.mockReturnValue(false)
-
-      mockKubectlProxy.exec.mockResolvedValue({
-        exitCode: 0,
-        output: JSON.stringify({
-          items: [{
-            metadata: { name: 'test-pod', namespace: 'default' },
-            spec: { hostNetwork: true, containers: [{ securityContext: { runAsNonRoot: true } }] },
-          }],
-        }),
-      })
-
-      const { useCachedSecurityIssues } = await loadModule()
-      useCachedSecurityIssues('prod')
-
-      const fetcher = capturedOpts.fetcher as () => Promise<Array<{ cluster: string }>>
-      const issues = await fetcher()
-
-      // Should only have scanned 'prod' cluster, not 'staging'
-      for (const issue of issues) {
-        expect(issue.cluster).toBe('prod')
-      }
-      // kubectlProxy.exec should have been called once (only for prod)
-      expect(mockKubectlProxy.exec).toHaveBeenCalledTimes(1)
-    })
-
-    it('security REST fallback returns empty on non-ok authFetch response', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      mockClusterCacheRef.clusters = [] as typeof mockClusterCacheRef.clusters
-      mockIsAgentUnavailable.mockReturnValue(true)
-      mockIsBackendUnavailable.mockReturnValue(false)
-
-      // authFetch returns non-ok
-      mockAuthFetch.mockResolvedValue({
-        ok: false,
-        status: 503,
-        json: vi.fn().mockResolvedValue(null),
-      })
-
-      const { useCachedSecurityIssues } = await loadModule()
-      useCachedSecurityIssues()
-
-      const fetcher = capturedOpts.fetcher as () => Promise<unknown>
-      // REST non-ok falls through to throw
-      await expect(fetcher()).rejects.toThrow('No data source available')
-    })
-
-    it('security REST fallback returns empty when authFetch JSON has empty issues', async () => {
-      let capturedOpts: Record<string, unknown> = {}
-      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
-        capturedOpts = opts
-        return makeCacheResult([])
-      })
-
-      mockClusterCacheRef.clusters = [] as typeof mockClusterCacheRef.clusters
-      mockIsAgentUnavailable.mockReturnValue(true)
-      mockIsBackendUnavailable.mockReturnValue(false)
-
-      // authFetch returns ok but with empty issues
-      mockAuthFetch.mockResolvedValue({
-        ok: true,
-        json: vi.fn().mockResolvedValue({ issues: [] }),
-      })
-
-      const { useCachedSecurityIssues } = await loadModule()
-      useCachedSecurityIssues()
-
-      const fetcher = capturedOpts.fetcher as () => Promise<unknown>
-      // Empty issues array doesn't satisfy `data.issues.length > 0`, falls through to throw
-      await expect(fetcher()).rejects.toThrow('No data source available')
-    })
-  })
 })
