+++ web/src/hooks/mcp/__tests__/config-fallback-modes.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
-import { renderHook, act, waitFor } from '@testing-library/react'
+import { renderHook, waitFor } from '@testing-library/react'
 
 // ---------------------------------------------------------------------------
 // Hoisted mocks
@@ -114,618 +114,6 @@ afterEach(() => {
 
 // ===========================================================================
 // useConfigMaps
-// ===========================================================================
-
-describe('useConfigMaps', () => {
-  it('returns empty array with loading state on mount', () => {
-    mockFetchSSE.mockReturnValue(new Promise(() => {}))
-    const { result } = renderHook(() => useConfigMaps())
-    expect(result.current.isLoading).toBe(true)
-    expect(result.current.configmaps).toEqual([])
-  })
-
-  it('returns config maps after SSE fetch resolves', async () => {
-    const fakeCMs = [{ name: 'cm-1', namespace: 'default', cluster: 'c1', dataCount: 2, age: '5d' }]
-    mockFetchSSE.mockResolvedValue(fakeCMs)
-
-    const { result } = renderHook(() => useConfigMaps())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.configmaps).toEqual(fakeCMs)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('forwards cluster and namespace via SSE params when provided', async () => {
-    mockFetchSSE.mockResolvedValue([])
-
-    renderHook(() => useConfigMaps('my-cluster', 'my-ns'))
-
-    await waitFor(() => expect(mockFetchSSE).toHaveBeenCalled())
-    const callArgs = mockFetchSSE.mock.calls[0][0] as { params: Record<string, string> }
-    expect(callArgs.params?.cluster).toBe('my-cluster')
-    expect(callArgs.params?.namespace).toBe('my-ns')
-  })
-
-  it('refetch() triggers a new SSE fetch', async () => {
-    mockFetchSSE.mockResolvedValue([])
-    const { result } = renderHook(() => useConfigMaps())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    const callsBefore = mockFetchSSE.mock.calls.length
-
-    await act(async () => { result.current.refetch() })
-
-    await waitFor(() => expect(mockFetchSSE.mock.calls.length).toBeGreaterThan(callsBefore))
-  })
-
-  it('re-fetches when demo mode changes', async () => {
-    mockFetchSSE.mockResolvedValue([])
-    const { result, rerender } = renderHook(
-      ({ demoMode }) => {
-        mockUseDemoMode.mockReturnValue({ isDemoMode: demoMode })
-        return useConfigMaps()
-      },
-      { initialProps: { demoMode: false } }
-    )
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    const callsBefore = mockFetchSSE.mock.calls.length
-
-    // Trigger demo mode change — hook registers an effect that calls refetch()
-    mockIsDemoMode.mockReturnValue(true)
-    rerender({ demoMode: true })
-
-    // In demo mode, refetch short-circuits before calling SSE, so configmaps should be demo data
-    await waitFor(() => expect(result.current.configmaps.length).toBeGreaterThan(0))
-    // Demo path bypasses SSE entirely — call count stays the same
-    expect(mockFetchSSE.mock.calls.length).toBe(callsBefore)
-  })
-
-  it('returns empty config maps with error: null on SSE and REST failure', async () => {
-    // Both SSE and REST fail — hook silently swallows error (configmaps are optional)
-    mockFetchSSE.mockRejectedValue(new Error('SSE error'))
-    globalThis.fetch = vi.fn().mockRejectedValue(new Error('REST error'))
-
-    const { result } = renderHook(() => useConfigMaps())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.configmaps).toEqual([])
-    expect(result.current.error).toBeNull()
-  })
-
-  it('returns demo config maps when demo mode is active', async () => {
-    mockIsDemoMode.mockReturnValue(true)
-    mockUseDemoMode.mockReturnValue({ isDemoMode: true })
-
-    const { result } = renderHook(() => useConfigMaps())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.configmaps.length).toBeGreaterThan(0)
-    expect(result.current.error).toBeNull()
-  })
-})
-
-// ===========================================================================
-// useSecrets
-// ===========================================================================
-
-describe('useSecrets', () => {
-  it('returns empty array with loading state on mount', () => {
-    mockFetchSSE.mockReturnValue(new Promise(() => {}))
-    const { result } = renderHook(() => useSecrets())
-    expect(result.current.isLoading).toBe(true)
-    expect(result.current.secrets).toEqual([])
-  })
-
-  it('returns secrets after SSE fetch resolves', async () => {
-    const fakeSecrets = [{ name: 'secret-1', namespace: 'default', cluster: 'c1', type: 'Opaque', dataCount: 3, age: '10d' }]
-    mockFetchSSE.mockResolvedValue(fakeSecrets)
-
-    const { result } = renderHook(() => useSecrets())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.secrets).toEqual(fakeSecrets)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('forwards cluster and namespace via SSE params when provided', async () => {
-    mockFetchSSE.mockResolvedValue([])
-
-    renderHook(() => useSecrets('cluster-x', 'ns-y'))
-
-    await waitFor(() => expect(mockFetchSSE).toHaveBeenCalled())
-    const callArgs = mockFetchSSE.mock.calls[0][0] as { params: Record<string, string> }
-    expect(callArgs.params?.cluster).toBe('cluster-x')
-    expect(callArgs.params?.namespace).toBe('ns-y')
-  })
-
-  it('refetch() triggers a new SSE fetch', async () => {
-    mockFetchSSE.mockResolvedValue([])
-    const { result } = renderHook(() => useSecrets())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    const callsBefore = mockFetchSSE.mock.calls.length
-
-    await act(async () => { result.current.refetch() })
-
-    await waitFor(() => expect(mockFetchSSE.mock.calls.length).toBeGreaterThan(callsBefore))
-  })
-
-  it('re-fetches when demo mode changes', async () => {
-    mockFetchSSE.mockResolvedValue([])
-    const { result, rerender } = renderHook(
-      ({ demoMode }) => {
-        mockUseDemoMode.mockReturnValue({ isDemoMode: demoMode })
-        return useSecrets()
-      },
-      { initialProps: { demoMode: false } }
-    )
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    const callsBefore = mockFetchSSE.mock.calls.length
-
-    // Trigger demo mode change — hook should re-fetch and return demo secrets
-    mockIsDemoMode.mockReturnValue(true)
-    rerender({ demoMode: true })
-
-    // In demo mode the hook short-circuits to demo data
-    await waitFor(() => expect(result.current.secrets.length).toBeGreaterThan(0))
-    // Demo path bypasses SSE entirely — call count stays the same
-    expect(mockFetchSSE.mock.calls.length).toBe(callsBefore)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('returns empty secrets with error: null on SSE and REST failure', async () => {
-    // Both SSE and REST fail — hook silently swallows error (secrets are optional)
-    mockFetchSSE.mockRejectedValue(new Error('SSE error'))
-    globalThis.fetch = vi.fn().mockRejectedValue(new Error('REST error'))
-
-    const { result } = renderHook(() => useSecrets())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.secrets).toEqual([])
-    expect(result.current.error).toBeNull()
-  })
-
-  it('returns demo secrets when demo mode is active', async () => {
-    mockIsDemoMode.mockReturnValue(true)
-    mockUseDemoMode.mockReturnValue({ isDemoMode: true })
-
-    const { result } = renderHook(() => useSecrets())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.secrets.length).toBeGreaterThan(0)
-    expect(result.current.error).toBeNull()
-  })
-})
-
-// ===========================================================================
-// useServiceAccounts
-// ===========================================================================
-
-describe('useServiceAccounts', () => {
-  it('returns empty array with loading state on mount', () => {
-    globalThis.fetch = vi.fn().mockImplementation(() => new Promise(() => {}))
-    const { result } = renderHook(() => useServiceAccounts())
-    expect(result.current.isLoading).toBe(true)
-    expect(result.current.serviceAccounts).toEqual([])
-  })
-
-  it('returns service accounts after REST fetch resolves', async () => {
-    const fakeSAs = [{ name: 'default', namespace: 'default', cluster: 'c1', secrets: ['default-token'], age: '30d' }]
-    globalThis.fetch = vi.fn().mockImplementation(() =>
-      Promise.resolve(new Response(JSON.stringify({ serviceAccounts: fakeSAs }), { status: 200 }))
-    )
-    // SSE fails to force the REST path
-    mockFetchSSE.mockRejectedValue(new Error('no SSE'))
-
-    const { result } = renderHook(() => useServiceAccounts())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.serviceAccounts).toEqual(fakeSAs)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('forwards cluster and namespace when provided', async () => {
-    globalThis.fetch = vi.fn().mockImplementation(() =>
-      Promise.resolve(new Response(JSON.stringify({ serviceAccounts: [] }), { status: 200 }))
-    )
-    mockFetchSSE.mockRejectedValue(new Error('no SSE'))
-
-    renderHook(() => useServiceAccounts('my-cluster', 'my-ns'))
-
-    await waitFor(() => expect(globalThis.fetch).toHaveBeenCalled())
-    const url: string = (globalThis.fetch as ReturnType<typeof vi.fn>).mock.calls[0][0]
-    expect(url).toContain('cluster=my-cluster')
-    expect(url).toContain('namespace=my-ns')
-  })
-
-  it('refetch() triggers a new fetch', async () => {
-    globalThis.fetch = vi.fn().mockImplementation(() =>
-      Promise.resolve(new Response(JSON.stringify({ serviceAccounts: [] }), { status: 200 }))
-    )
-    const { result } = renderHook(() => useServiceAccounts())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    const callsBefore = (globalThis.fetch as ReturnType<typeof vi.fn>).mock.calls.length
-
-    await act(async () => { result.current.refetch() })
-
-    await waitFor(() => expect((globalThis.fetch as ReturnType<typeof vi.fn>).mock.calls.length).toBeGreaterThan(callsBefore))
-  })
-
-  it('returns empty service accounts with error: null on failure', async () => {
-    // Both SSE and REST fail — hook silently swallows error (service accounts are optional)
-    mockFetchSSE.mockRejectedValue(new Error('SSE error'))
-    globalThis.fetch = vi.fn().mockRejectedValue(new Error('REST error'))
-
-    const { result } = renderHook(() => useServiceAccounts())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.serviceAccounts).toEqual([])
-    expect(result.current.error).toBeNull()
-  })
-
-  it('returns demo service accounts when demo mode is active', async () => {
-    mockIsDemoMode.mockReturnValue(true)
-    mockUseDemoMode.mockReturnValue({ isDemoMode: true })
-
-    const { result } = renderHook(() => useServiceAccounts())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.serviceAccounts.length).toBeGreaterThan(0)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('re-fetches when demo mode changes', async () => {
-    globalThis.fetch = vi.fn().mockImplementation(() =>
-      Promise.resolve(new Response(JSON.stringify({ serviceAccounts: [] }), { status: 200 }))
-    )
-    const { result, rerender } = renderHook(
-      ({ demoMode }) => {
-        mockUseDemoMode.mockReturnValue({ isDemoMode: demoMode })
-        return useServiceAccounts()
-      },
-      { initialProps: { demoMode: false } }
-    )
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Trigger demo mode change
-    mockIsDemoMode.mockReturnValue(true)
-    rerender({ demoMode: true })
-
-    await waitFor(() => expect(result.current.serviceAccounts.length).toBeGreaterThan(0))
-    expect(result.current.error).toBeNull()
-  })
-})
-
-// ===========================================================================
-// Regression tests: local agent path
-// ===========================================================================
-
-describe('useConfigMaps — local agent path', () => {
-  it('fetches from local agent when cluster is provided and agent is available', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    const agentCMs = [{ name: 'agent-cm', namespace: 'ns1', cluster: 'c1', dataCount: 1, age: '1d' }]
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: () => Promise.resolve({ configmaps: agentCMs }),
-    })
-
-    const { result } = renderHook(() => useConfigMaps('c1', 'ns1'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.configmaps).toEqual(agentCMs)
-    expect(result.current.error).toBeNull()
-    expect(mockReportAgentDataSuccess).toHaveBeenCalled()
-    // SSE and REST should NOT have been called
-    expect(mockFetchSSE).not.toHaveBeenCalled()
-    expect(mockApiGet).not.toHaveBeenCalled()
-  })
-
-  it('falls through to SSE when local agent returns non-ok response', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    globalThis.fetch = vi.fn().mockResolvedValue({ ok: false, status: 500 })
-    const sseCMs = [{ name: 'sse-cm', namespace: 'ns1', cluster: 'c1', dataCount: 2, age: '3d' }]
-    mockFetchSSE.mockResolvedValue(sseCMs)
-
-    const { result } = renderHook(() => useConfigMaps('c1', 'ns1'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.configmaps).toEqual(sseCMs)
-    expect(mockReportAgentDataSuccess).not.toHaveBeenCalled()
-  })
-
-  it('falls through to SSE when local agent fetch throws', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    globalThis.fetch = vi.fn().mockRejectedValue(new Error('network error'))
-    mockFetchSSE.mockResolvedValue([])
-
-    const { result } = renderHook(() => useConfigMaps('c1'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.configmaps).toEqual([])
-    expect(mockFetchSSE).toHaveBeenCalled()
-  })
-
-  it('skips local agent when cluster is not provided', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    globalThis.fetch = vi.fn()
-    mockFetchSSE.mockResolvedValue([])
-
-    const { result } = renderHook(() => useConfigMaps())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    // globalThis.fetch should NOT have been called (local agent path requires cluster)
-    expect(globalThis.fetch).not.toHaveBeenCalled()
-    expect(mockFetchSSE).toHaveBeenCalled()
-  })
-
-  it('handles local agent returning empty configmaps array', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: () => Promise.resolve({ configmaps: [] }),
-    })
-
-    const { result } = renderHook(() => useConfigMaps('c1'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.configmaps).toEqual([])
-    expect(result.current.error).toBeNull()
-    expect(mockReportAgentDataSuccess).toHaveBeenCalled()
-  })
-
-  it('handles local agent returning response without configmaps key (defaults to [])', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: () => Promise.resolve({}),
-    })
-
-    const { result } = renderHook(() => useConfigMaps('c1'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.configmaps).toEqual([])
-    expect(mockReportAgentDataSuccess).toHaveBeenCalled()
-  })
-
-  it('appends namespace to local agent URL when provided', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: () => Promise.resolve({ configmaps: [] }),
-    })
-
-    renderHook(() => useConfigMaps('c1', 'my-ns'))
-
-    await waitFor(() => expect(globalThis.fetch).toHaveBeenCalled())
-    const fetchUrl = (globalThis.fetch as ReturnType<typeof vi.fn>).mock.calls[0][0] as string
-    expect(fetchUrl).toContain('cluster=c1')
-    expect(fetchUrl).toContain('namespace=my-ns')
-  })
-})
-
-describe('useSecrets — local agent path', () => {
-  it('fetches from local agent when cluster is provided and agent is available', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    const agentSecrets = [
-      { name: 'tls-cert', namespace: 'default', cluster: 'c1', type: 'kubernetes.io/tls', dataCount: 2, age: '5d' },
-    ]
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: () => Promise.resolve({ secrets: agentSecrets }),
-    })
-
-    const { result } = renderHook(() => useSecrets('c1', 'default'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.secrets).toEqual(agentSecrets)
-    expect(result.current.error).toBeNull()
-    expect(mockReportAgentDataSuccess).toHaveBeenCalled()
-  })
-
-  it('falls through to SSE when local agent returns non-ok', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    globalThis.fetch = vi.fn().mockResolvedValue({ ok: false, status: 403 })
-    const sseSecrets = [
-      { name: 'sse-secret', namespace: 'ns', cluster: 'c1', type: 'Opaque', dataCount: 1, age: '1d' },
-    ]
-    mockFetchSSE.mockResolvedValue(sseSecrets)
-
-    const { result } = renderHook(() => useSecrets('c1'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.secrets).toEqual(sseSecrets)
-  })
-
-  it('handles local agent returning response without secrets key', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: () => Promise.resolve({}),
-    })
-
-    const { result } = renderHook(() => useSecrets('c1'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.secrets).toEqual([])
-  })
-
-  it('falls through to SSE when local agent fetch throws', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    globalThis.fetch = vi.fn().mockRejectedValue(new Error('agent down'))
-    mockFetchSSE.mockResolvedValue([])
-
-    const { result } = renderHook(() => useSecrets('c1'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.secrets).toEqual([])
-    expect(mockFetchSSE).toHaveBeenCalled()
-  })
-
-  it('appends namespace to local agent URL when provided', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: () => Promise.resolve({ secrets: [] }),
-    })
-
-    renderHook(() => useSecrets('c1', 'kube-system'))
-
-    await waitFor(() => expect(globalThis.fetch).toHaveBeenCalled())
-    const fetchUrl = (globalThis.fetch as ReturnType<typeof vi.fn>).mock.calls[0][0] as string
-    expect(fetchUrl).toContain('cluster=c1')
-    expect(fetchUrl).toContain('namespace=kube-system')
-  })
-})
-
-// ===========================================================================
-// Regression tests: SSE streaming behavior
-// ===========================================================================
-
-describe('useConfigMaps — SSE streaming', () => {
-  it('uses correct SSE URL and itemsKey', async () => {
-    mockFetchSSE.mockResolvedValue([])
-
-    renderHook(() => useConfigMaps('c1', 'ns1'))
-
-    await waitFor(() => expect(mockFetchSSE).toHaveBeenCalled())
-    const sseArg = mockFetchSSE.mock.calls[0][0] as {
-      url: string
-      itemsKey: string
-      params: Record<string, string>
-    }
-    expect(sseArg.url).toBe(`${LOCAL_AGENT_HTTP_URL}/configmaps/stream`)
-    expect(sseArg.itemsKey).toBe('configmaps')
-  })
-
-  it('omits cluster/namespace from SSE params when not provided', async () => {
-    mockFetchSSE.mockResolvedValue([])
-
-    renderHook(() => useConfigMaps())
-
-    await waitFor(() => expect(mockFetchSSE).toHaveBeenCalled())
-    const sseArg = mockFetchSSE.mock.calls[0][0] as { params: Record<string, string> }
-    expect(sseArg.params.cluster).toBeUndefined()
-    expect(sseArg.params.namespace).toBeUndefined()
-  })
-
-  it('skips SSE when no token is present and falls through to REST', async () => {
-    localStorage.removeItem('token')
-    const restCMs = [{ name: 'rest-cm', namespace: 'default', cluster: 'c1', dataCount: 1, age: '1d' }]
-    globalThis.fetch = vi.fn().mockImplementation(() =>
-      Promise.resolve(new Response(JSON.stringify({ configmaps: restCMs }), { status: 200 }))
-    )
-
-    const { result } = renderHook(() => useConfigMaps())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(mockFetchSSE).not.toHaveBeenCalled()
-    expect(result.current.configmaps).toEqual(restCMs)
-  })
-
-  it('skips SSE when token is demo-token and falls through to REST', async () => {
-    localStorage.setItem('token', 'demo-token')
-    const restCMs = [{ name: 'rest-cm', namespace: 'default', cluster: 'c1', dataCount: 1, age: '1d' }]
-    globalThis.fetch = vi.fn().mockImplementation(() =>
-      Promise.resolve(new Response(JSON.stringify({ configmaps: restCMs }), { status: 200 }))
-    )
-
-    const { result } = renderHook(() => useConfigMaps())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(mockFetchSSE).not.toHaveBeenCalled()
-    expect(result.current.configmaps).toEqual(restCMs)
-  })
-
-  it('invokes onClusterData callback during SSE streaming for configmaps', async () => {
-    const streamedItems = [
-      { name: 'cm-a', namespace: 'ns1', cluster: 'c1', dataCount: 1, age: '1d' },
-      { name: 'cm-b', namespace: 'ns2', cluster: 'c2', dataCount: 2, age: '2d' },
-    ]
-    // Simulate fetchSSE calling onClusterData before resolving
-    mockFetchSSE.mockImplementation(async (opts: { onClusterData?: (cluster: string, items: unknown[]) => void }) => {
-      if (opts.onClusterData) {
-        opts.onClusterData('c1', [streamedItems[0]])
-        opts.onClusterData('c2', [streamedItems[1]])
-      }
-      return streamedItems
-    })
-
-    const { result } = renderHook(() => useConfigMaps())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.configmaps).toEqual(streamedItems)
-  })
-})
-
-describe('useSecrets — SSE streaming', () => {
-  it('uses correct SSE URL and itemsKey', async () => {
-    mockFetchSSE.mockResolvedValue([])
-
-    renderHook(() => useSecrets('c1'))
-
-    await waitFor(() => expect(mockFetchSSE).toHaveBeenCalled())
-    const sseArg = mockFetchSSE.mock.calls[0][0] as {
-      url: string
-      itemsKey: string
-    }
-    expect(sseArg.url).toBe(`${LOCAL_AGENT_HTTP_URL}/secrets/stream`)
-    expect(sseArg.itemsKey).toBe('secrets')
-  })
-
-  it('invokes onClusterData callback during SSE streaming for secrets', async () => {
-    const streamedSecrets = [
-      { name: 'secret-a', namespace: 'ns1', cluster: 'c1', type: 'Opaque', dataCount: 1, age: '1d' },
-      { name: 'secret-b', namespace: 'ns2', cluster: 'c2', type: 'Opaque', dataCount: 2, age: '2d' },
-    ]
-    mockFetchSSE.mockImplementation(async (opts: { onClusterData?: (cluster: string, items: unknown[]) => void }) => {
-      if (opts.onClusterData) {
-        opts.onClusterData('c1', [streamedSecrets[0]])
-        opts.onClusterData('c2', [streamedSecrets[1]])
-      }
-      return streamedSecrets
-    })
-
-    const { result } = renderHook(() => useSecrets())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.secrets).toEqual(streamedSecrets)
-  })
-
-  it('skips SSE when no token is present and falls through to REST for secrets', async () => {
-    localStorage.removeItem('token')
-    const restSecrets = [{ name: 'rest-s', namespace: 'default', cluster: 'c1', type: 'Opaque', dataCount: 1, age: '1d' }]
-    globalThis.fetch = vi.fn().mockImplementation(() =>
-      Promise.resolve(new Response(JSON.stringify({ secrets: restSecrets }), { status: 200 }))
-    )
-
-    const { result } = renderHook(() => useSecrets())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(mockFetchSSE).not.toHaveBeenCalled()
-    expect(result.current.secrets).toEqual(restSecrets)
-  })
-
-  it('skips SSE when token is demo-token and falls through to REST for secrets', async () => {
-    localStorage.setItem('token', 'demo-token')
-    const restSecrets = [{ name: 'rest-s', namespace: 'default', cluster: 'c1', type: 'Opaque', dataCount: 1, age: '1d' }]
-    globalThis.fetch = vi.fn().mockImplementation(() =>
-      Promise.resolve(new Response(JSON.stringify({ secrets: restSecrets }), { status: 200 }))
-    )
-
-    const { result } = renderHook(() => useSecrets())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(mockFetchSSE).not.toHaveBeenCalled()
-    expect(result.current.secrets).toEqual(restSecrets)
-  })
-})
-
-// ===========================================================================
-// Regression tests: REST fallback
 // ===========================================================================
 
 describe('useConfigMaps — REST fallback', () => {
